Updated program to store the current commission in the manager account, to reduce dependency on vote account data storage format.  Modified the enter instruction to fail if the vote account is of an unknown data storage format.
Modified leave to deallocate the data of the manager account so
that it cannot be reconstituted by a subsequent instruction in the same
express (via lamports transfer) and then end up preventing re-enter of the
program.

Added a test to ensure that the deallocate of data has the expected effect.